+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-styles/src/main/java/com/liferay/fragment/entry/processor/styles/StylesFragmentEntryProcessor.java
@@ -17,26 +17,15 @@ package com.liferay.fragment.entry.processor.styles;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.processor.FragmentEntryProcessor;
 import com.liferay.fragment.processor.FragmentEntryProcessorContext;
-import com.liferay.layout.constants.LayoutWebKeys;
-import com.liferay.layout.page.template.model.LayoutPageTemplateStructure;
-import com.liferay.layout.page.template.service.LayoutPageTemplateStructureLocalService;
-import com.liferay.layout.util.structure.FragmentStyledLayoutStructureItem;
-import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-
-import javax.servlet.http.HttpServletRequest;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Víctor Galán
@@ -72,40 +61,7 @@ public class StylesFragmentEntryProcessor implements FragmentEntryProcessor {
 		FragmentEntryLink fragmentEntryLink, String html,
 		FragmentEntryProcessorContext fragmentEntryProcessorContext) {
 
-		Document document = _getDocument(html);
-
-		Elements elements = document.select("[data-lfr-styles]");
-
-		if (elements.isEmpty()) {
-			return html;
-		}
-
-		FragmentStyledLayoutStructureItem fragmentStyledLayoutStructureItem =
-			_getLayoutStructureItem(
-				fragmentEntryLink,
-				fragmentEntryProcessorContext.getHttpServletRequest());
-
-		if (fragmentStyledLayoutStructureItem == null) {
-			return html;
-		}
-
-		String fragmentEntryLinkCssClass =
-			fragmentStyledLayoutStructureItem.getFragmentEntryLinkCssClass(
-				fragmentEntryLink);
-		String layoutStructureItemUniqueCssClass =
-			fragmentStyledLayoutStructureItem.getUniqueCssClass();
-		String styledLayoutStructureItemCssClasses =
-			fragmentStyledLayoutStructureItem.getStyledCssClasses();
-
-		for (Element element : elements) {
-			element.addClass(fragmentEntryLinkCssClass);
-			element.addClass(layoutStructureItemUniqueCssClass);
-			element.addClass(styledLayoutStructureItemCssClasses);
-		}
-
-		Element bodyElement = document.body();
-
-		return bodyElement.html();
+		return html;
 	}
 
 	private Document _getDocument(String html) {
@@ -119,49 +75,5 @@ public class StylesFragmentEntryProcessor implements FragmentEntryProcessor {
 
 		return document;
 	}
-
-	private FragmentStyledLayoutStructureItem _getLayoutStructureItem(
-		FragmentEntryLink fragmentEntryLink,
-		HttpServletRequest httpServletRequest) {
-
-		LayoutStructure layoutStructure = null;
-
-		if (httpServletRequest != null) {
-			layoutStructure = (LayoutStructure)httpServletRequest.getAttribute(
-				LayoutWebKeys.LAYOUT_STRUCTURE);
-		}
-
-		if (layoutStructure == null) {
-			try {
-				LayoutPageTemplateStructure layoutPageTemplateStructure =
-					_layoutPageTemplateStructureLocalService.
-						fetchLayoutPageTemplateStructure(
-							fragmentEntryLink.getGroupId(),
-							fragmentEntryLink.getPlid(), true);
-
-				layoutStructure = LayoutStructure.of(
-					layoutPageTemplateStructure.getData(
-						fragmentEntryLink.getSegmentsExperienceId()));
-			}
-			catch (Exception exception) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(exception);
-				}
-
-				return null;
-			}
-		}
-
-		return (FragmentStyledLayoutStructureItem)
-			layoutStructure.getLayoutStructureItemByFragmentEntryLinkId(
-				fragmentEntryLink.getFragmentEntryLinkId());
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		StylesFragmentEntryProcessor.class);
-
-	@Reference
-	private LayoutPageTemplateStructureLocalService
-		_layoutPageTemplateStructureLocalService;
 
 }
